+++ common/src/comp/inventory/item/mod.rs
@@ -368,7 +368,7 @@ pub enum ItemKind {
     Ingredient {
         /// Used to generate names for modular items composed of this ingredient
         // I think we can actually remove it now?
-        #[deprecated = "part of non-localized name generation"]
+        #[deprecated = "since item i18n"]
         descriptor: String,
     },
     TagExamples {
@@ -556,6 +556,18 @@ type I18nId = String;
 pub struct ItemI18n {
     /// maps ItemKey to i18n identifier
     map: HashMap<ItemKey, I18nId>,
+    /// maps FragmentKey to i18n identifier
+    ///
+    /// Used for optional templating for languages that can stomach them
+    fragments: HashMap<FragmentKey, I18nId>,
+}
+
+#[derive(Hash, Eq, PartialEq, Debug, Clone, Deserialize, Serialize)]
+pub enum FragmentKey {
+    // path to ingredient
+    Ingredient(String),
+    // path to primary component and hand-ness required
+    WeaponPrimaryComponent(String, Hands),
 }
 
 impl ItemI18n {
@@ -569,19 +581,35 @@ impl ItemI18n {
     /// Returns (name, description) in Content form.
     // TODO: after we remove legacy text from ItemDef, consider making this
     // function non-fallible?
-    fn item_text_opt(&self, mut item_key: ItemKey) -> Option<(Content, Content)> {
-        // we don't put TagExamples into manifest
-        if let ItemKey::TagExamples(_, id) = item_key {
-            item_key = ItemKey::Simple(id.to_string());
-        }
+    fn item_text_opt(&self, item_key: &ItemKey) -> Option<(Content, Content)> {
+        // We don't put TagExamples into manifest.
+        // Instead they are marked as Simple.
+        let key;
+        let item_key = if let ItemKey::TagExamples(_, id) = item_key {
+            key = ItemKey::Simple(id.to_string());
+            &key
+        } else {
+            item_key
+        };
 
-        let key = self.map.get(&item_key);
+        let key = self.map.get(item_key);
         key.map(|key| {
             (
                 Content::Key(key.to_owned()),
                 Content::Attr(key.to_owned(), "desc".to_owned()),
             )
         })
+    }
+
+    fn try_fragment(&self, fragment_key: &FragmentKey) -> Option<Content> {
+        self.fragments
+            .get(fragment_key)
+            .map(|key| Content::Key(key.to_owned()))
+    }
+
+    /// Returns all fragments, mainly for testing
+    pub fn all_fragments(&self) -> impl Iterator<Item = (&FragmentKey, &I18nId)> {
+        self.fragments.iter()
     }
 }
 
@@ -751,9 +779,7 @@ pub struct ItemDef {
     /// prepended with `veloren.core` is reserved for veloren functions.
     item_definition_id: String,
     #[deprecated = "since item i18n"]
-    name: String,
-    #[deprecated = "since item i18n"]
-    description: String,
+    legacy_name: String,
     pub kind: ItemKind,
     pub quality: Quality,
     pub tags: Vec<ItemTag>,
@@ -863,8 +889,7 @@ impl ItemDef {
         #[expect(deprecated)]
         Self {
             item_definition_id,
-            name: "test item name".to_owned(),
-            description: "test item description".to_owned(),
+            legacy_name: "test item name".to_owned(),
             kind,
             quality,
             tags,
@@ -878,8 +903,7 @@ impl ItemDef {
         #[expect(deprecated)]
         Self {
             item_definition_id: "test.item".to_string(),
-            name: "test item name".to_owned(),
-            description: "test item description".to_owned(),
+            legacy_name: "test item name".to_owned(),
             kind,
             quality: Quality::Common,
             tags: vec![],
@@ -920,7 +944,7 @@ impl Asset for ItemDef {
 
         let RawItemDef {
             legacy_name,
-            legacy_description,
+            legacy_description: _,
             kind,
             quality,
             tags,
@@ -934,11 +958,10 @@ impl Asset for ItemDef {
         // TODO: This probably does not belong here
         let item_definition_id = specifier.replace('\\', ".");
 
-        #[expect(deprecated)]
         Ok(ItemDef {
             item_definition_id,
-            name: legacy_name,
-            description: legacy_description,
+            #[expect(deprecated)]
+            legacy_name,
             kind,
             quality,
             tags,
@@ -1296,28 +1319,19 @@ impl Item {
     }
 
     #[deprecated = "since item i18n"]
-    pub fn name(&self) -> Cow<'_, str> {
+    pub fn legacy_name(&self) -> Cow<'_, str> {
         match &self.item_base {
             ItemBase::Simple(item_def) => {
                 if self.components.is_empty() {
                     #[expect(deprecated)]
-                    Cow::Borrowed(&item_def.name)
+                    Cow::Borrowed(&item_def.legacy_name)
                 } else {
                     #[expect(deprecated)]
-                    modular::modify_name(&item_def.name, self)
+                    modular::modify_name(&item_def.legacy_name, self)
                 }
             },
+            #[expect(deprecated, reason = "since item i18n")]
             ItemBase::Modular(mod_base) => mod_base.generate_name(self.components()),
-        }
-    }
-
-    #[deprecated = "since item i18n"]
-    pub fn description(&self) -> &str {
-        match &self.item_base {
-            #[expect(deprecated)]
-            ItemBase::Simple(item_def) => &item_def.description,
-            // TODO: See if James wanted to make description, else leave with none
-            ItemBase::Modular(_) => "",
         }
     }
 
@@ -1746,9 +1760,7 @@ pub fn flatten_counted_items<'a>(
 /// for either an `Item` containing the definition, or the actual `ItemDef`
 pub trait ItemDesc {
     #[deprecated = "since item i18n"]
-    fn description(&self) -> &str;
-    #[deprecated = "since item i18n"]
-    fn name(&self) -> Cow<'_, str>;
+    fn legacy_name(&self) -> Cow<'_, str>;
     fn kind(&self) -> Cow<'_, ItemKind>;
     fn amount(&self) -> NonZeroU32;
     fn quality(&self) -> Quality;
@@ -1773,25 +1785,49 @@ pub trait ItemDesc {
     fn i18n(&self, i18n: &ItemI18n) -> (Content, Content) {
         let item_key: ItemKey = self.into();
 
-        #[expect(deprecated)]
-        i18n.item_text_opt(item_key).unwrap_or_else(|| {
+        let (name, description) = i18n.item_text_opt(&item_key).unwrap_or_else(|| {
             (
-                Content::Plain(self.name().to_string()),
-                Content::Plain(self.description().to_string()),
+                #[expect(deprecated)]
+                Content::Plain(self.legacy_name().to_string()),
+                Content::Plain(String::new()),
             )
-        })
+        });
+
+        if let ItemKey::ModularWeapon((comp_id, ing_id, hands)) = item_key {
+            let b = |x| Box::new(x);
+
+            // the name template
+            let title_fallback = Content::localized("weapon-modular-fallback-template")
+                .with_arg(
+                    "material-fragment",
+                    i18n.try_fragment(&FragmentKey::Ingredient(ing_id))
+                        // use Key instead of Plain here, so it's marked as
+                        // "dirty" during and attempts English
+                        .unwrap_or_else(|| Content::Key("Modular".to_owned())),
+                )
+                .with_arg(
+                    "weapon",
+                    i18n.try_fragment(&FragmentKey::WeaponPrimaryComponent(comp_id, hands))
+                        // use Key instead of Plain here, so it's marked as
+                        // "dirty" during and attempts English
+                        .unwrap_or_else(|| Content::Key("Weapon".to_owned())),
+                );
+
+            (
+                Content::WithFallback(b(name), b(title_fallback)),
+                // no fallback for description
+                description,
+            )
+        } else {
+            (name, description)
+        }
     }
 }
 
 impl ItemDesc for Item {
-    fn description(&self) -> &str {
+    fn legacy_name(&self) -> Cow<'_, str> {
         #[expect(deprecated)]
-        self.description()
-    }
-
-    fn name(&self) -> Cow<'_, str> {
-        #[expect(deprecated)]
-        self.name()
+        self.legacy_name()
     }
 
     fn kind(&self) -> Cow<'_, ItemKind> { self.kind() }
@@ -1820,14 +1856,9 @@ impl ItemDesc for Item {
 }
 
 impl ItemDesc for FrontendItem {
-    fn description(&self) -> &str {
+    fn legacy_name(&self) -> Cow<'_, str> {
         #[expect(deprecated)]
-        self.0.description()
-    }
-
-    fn name(&self) -> Cow<'_, str> {
-        #[expect(deprecated)]
-        self.0.name()
+        self.0.legacy_name()
     }
 
     fn kind(&self) -> Cow<'_, ItemKind> { self.0.kind() }
@@ -1856,14 +1887,9 @@ impl ItemDesc for FrontendItem {
 }
 
 impl ItemDesc for ItemDef {
-    fn description(&self) -> &str {
+    fn legacy_name(&self) -> Cow<'_, str> {
         #[expect(deprecated)]
-        &self.description
-    }
-
-    fn name(&self) -> Cow<'_, str> {
-        #[expect(deprecated)]
-        Cow::Borrowed(&self.name)
+        Cow::Borrowed(&self.legacy_name)
     }
 
     fn kind(&self) -> Cow<'_, ItemKind> { Cow::Borrowed(&self.kind) }
@@ -1894,14 +1920,9 @@ impl ItemDesc for ItemDef {
 }
 
 impl ItemDesc for PickupItem {
-    fn description(&self) -> &str {
+    fn legacy_name(&self) -> Cow<'_, str> {
         #[expect(deprecated)]
-        self.item().description()
-    }
-
-    fn name(&self) -> Cow<'_, str> {
-        #[expect(deprecated)]
-        self.item().name()
+        self.item().legacy_name()
     }
 
     fn kind(&self) -> Cow<'_, ItemKind> { self.item().kind() }
@@ -1950,14 +1971,9 @@ impl Component for ThrownItem {
 pub struct DurabilityMultiplier(pub f32);
 
 impl<T: ItemDesc + ?Sized> ItemDesc for &T {
-    fn description(&self) -> &str {
+    fn legacy_name(&self) -> Cow<'_, str> {
         #[expect(deprecated)]
-        (*self).description()
-    }
-
-    fn name(&self) -> Cow<'_, str> {
-        #[expect(deprecated)]
-        (*self).name()
+        (*self).legacy_name()
     }
 
     fn kind(&self) -> Cow<'_, ItemKind> { (*self).kind() }
@@ -2100,6 +2116,7 @@ impl From<&ItemDefinitionId<'_>> for ItemDefinitionIdOwned {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use hashbrown::HashSet;
 
     #[test]
     fn test_assets_items() {
@@ -2137,12 +2154,42 @@ mod tests {
         let mut errs = vec![];
         for item in items {
             let item_key: ItemKey = (&item).into();
-            if manifest.item_text_opt(item_key.clone()).is_none() {
+            if manifest.item_text_opt(&item_key.clone()).is_none() {
                 errs.push(item_key)
             }
         }
         if !errs.is_empty() {
             panic!("item i18n manifest misses translation-id for following items {errs:#?}")
+        }
+    }
+
+    #[test]
+    // This exists to make translators' lives easier when translating
+    // modulars.
+    fn ensure_modular_fragments() {
+        let manifest = ItemI18n::new_expect();
+        let items = all_items_expect();
+        let mut errs = HashSet::new();
+
+        for item in items {
+            let item_key: ItemKey = (&item).into();
+            if let ItemKey::ModularWeapon((comp_id, ing_id, hands)) = item_key {
+                if manifest
+                    .try_fragment(&FragmentKey::Ingredient(ing_id.clone()))
+                    .is_none()
+                {
+                    errs.insert(FragmentKey::Ingredient(ing_id));
+                }
+                if manifest
+                    .try_fragment(&FragmentKey::WeaponPrimaryComponent(comp_id.clone(), hands))
+                    .is_none()
+                {
+                    errs.insert(FragmentKey::WeaponPrimaryComponent(comp_id, hands));
+                }
+            }
+        }
+        if !errs.is_empty() {
+            panic!("item i18n manifest missing fragment-id for following items {errs:#?}")
         }
     }
 }